chore: update dev dependencies, nuxt content compatibility, move cli outside of runtime folder
