+++ package.json
@@ -18,7 +18,7 @@
     "dist"
   ],
   "scripts": {
-    "prepack": "nuxt-module-build",
+    "prepack": "nuxt-module-build build",
     "dev": "nuxi dev playground",
     "dev:build": "nuxi build playground",
     "dev:generate": "nuxi generate playground",
@@ -35,22 +35,26 @@
     "vue-component-meta": "^1.8.22"
   },
   "devDependencies": {
-    "@iconify/vue": "^4.1.0",
-    "@nuxt/content": "^2.5.2",
-    "@nuxt/module-builder": "latest",
-    "@nuxt/test-utils": "^3.3.1",
-    "@nuxtjs/eslint-config-typescript": "latest",
-    "eslint": "latest",
-    "nuxt": "^3.3.1",
-    "release-it": "^15.9.1",
-    "vitest": "^0.29.7",
-    "vue": "^3.2.47"
+    "@iconify/vue": "^4.1.1",
+    "@nuxt/content": "^2.9.0",
+    "@nuxt/module-builder": "^0.5.4",
+    "@nuxt/test-utils": "^3.8.1",
+    "@nuxtjs/eslint-config-typescript": "^12.1.0",
+    "eslint": "^8.53.0",
+    "nuxt": "^3.8.1",
+    "release-it": "^17.0.0",
+    "vitest": "^0.34.6",
+    "vue": "^3.3.8"
   },
   "build": {
     "entries": [
       {
         "input": "./src/parser.ts",
         "name": "parser"
+      },
+      {
+        "input": "./src/cli/index.ts",
+        "name": "cli"
       }
     ],
     "externals": [
@@ -59,7 +63,11 @@
       "pathe",
       "defu",
       "unplugin",
-      "consola"
+      "consola",
+      "mlly",
+      "acorn",
+      "pkg-types",
+      "jsonc-parser"
     ]
   },
   "pnpm": {
