+++ playground/nuxt.config.ts
@@ -25,9 +25,7 @@ export default defineNuxtConfig({
 
   componentMeta: {
     debug: 2,
-    exclude: [/test/i, (component: any) => {
-      return component.global
-    }],
+    exclude: [/node_modules/i],
     metaSources: [
       {
         TestExternalMeta: {
